Move Util.cs into Utils folder
